requirements.txt
numpy == 1.19.1
pandas == 1.0.5
pandas-profiling == 2.8.0
plotly == 4.9.0
psutil == 5.7.2
py-cpuinfo == 7.0.0
scikit-learn == 0.23.1
scipy == 1.5.2
colorama == 0.4.4